week 3 report 17
DONE
tried the last test with the repository and got the same result.

PROBLEMS
the autowired annotation has a warning that says that neither the UserRepository nor the UserService are valid spring beans despite the former inheriting a repository interface and the later being annotated with @Service. dont know what that's about but it might be the source of my problems.